deinline fmpz IO
